refactor: 代码优化

+++ sharp-database2/src/main/java/com/rick/db/repository/support/IdToEntityConverterFactory.java
@@ -37,7 +37,7 @@ final public class IdToEntityConverterFactory implements ConverterFactory<Object
 
                 T t = (T) targetType.newInstance();
                 if (obj.getClass() == Long.class) {
-                    t.setId((Long) obj);
+                    t.setId(obj);
                 } else if (obj.getClass() == String.class) {
                     t.setId(Long.parseLong((String) obj));
                 }
